remove space_shooter.rs and create states module
