+++ src/states/mod.rs
@@ -1,0 +1,7 @@
+mod main_game;
+mod paused;
+
+pub use self::{
+    main_game::{MainGameState, TrackedStats},
+    paused::PausedState,
+};
